backend/requirements.txt
redis
shortuuid
dnspython
python-dotenv
pyopenssl
python-jose[cryptography]
passlib[bcrypt]
websockets
supervisor
asyncio
aioredis